Add customserializer, json output from chapter 13
I don't know why, but json.dump converts a time.struct_time object into
a list of integers instead of throwing a TypeError exception as
described in Dive Into. The json module documentation doesn't mention
anything about handling time.struct_time.

On the other hand, this has revealed to me some of the internal behavior
of json.dump. Clearly, it attempts to serialize an object on its own
first, and only then calls the function passed in the "default"
parameter.